requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
pgvector
openai
langchain
psycopg2-binary
chardet
langchain-openai
sentence-transformers
scikit-learn
matplotlib
seaborn
plotly
kaleido
fastapi
uvicorn[standard]
streamlit
chardet
python-multipart
requests
tqdm
google-generativeai
google